requirements.txt
sqlalchemy
dotenv
aiohttp
aiofiles
pillow